reorganized react dev app
