+++ src/react/.eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": "react-app",
+  "rules": {
+    "jsx-filename-extension": "off"
+  }
+}
